refactor(macros): expose `stmt::Json<T>` for deferred JSON lazy load
Move the hidden `LoadJson<T>` adapter from `schema/deferred.rs` to
`stmt/json.rs` and re-export it as `toasty::stmt::Json` — mirroring
the existing `stmt::List<T>` marker. The lazy-load accessor now
returns `Statement<stmt::Json<T>>`, which reads cleaner next to
`Statement<List<M>>` and lives in the right module for a `Statement`
type-parameter marker.
