+++ crates/toasty/src/schema.rs
@@ -9,8 +9,6 @@ pub mod create_meta;
 pub use create_meta::{CreateField, CreateMeta, ValidateCreate};
 
 mod deferred;
-#[cfg(feature = "serde")]
-pub use deferred::LoadJson;
 pub use deferred::{Defer, Deferred, build_deferred_load};
 
 mod embed;
