+++ crates/toasty/src/schema/deferred.rs
@@ -234,40 +234,3 @@ impl<T: serde_core::Serialize> serde_core::Serialize for Deferred<T> {
         self.value.serialize(serializer)
     }
 }
-
-/// `Load` adapter that decodes a JSON-encoded column into `T`.
-///
-/// Used by the per-field lazy-load accessor generated for fields annotated
-/// with both `#[serialize(json)]` and `#[deferred]`. The accessor returns
-/// `Statement<LoadJson<T>>`, whose `Output = T`, so `.exec()` yields the
-/// deserialized value directly.
-///
-/// `T` may itself be `Option<U>`; the impl maps a `Null` column to the JSON
-/// literal `null` before deserializing, which `Option<U>` decodes as `None`.
-/// Non-`Option` `T` instead surface a deserialization error on `Null`, which
-/// is the right behavior for a non-nullable column.
-#[doc(hidden)]
-#[cfg(feature = "serde")]
-pub struct LoadJson<T>(std::marker::PhantomData<T>);
-
-#[cfg(feature = "serde")]
-impl<T> Load for LoadJson<T>
-where
-    T: for<'de> serde_core::Deserialize<'de>,
-{
-    type Output = T;
-
-    fn ty() -> toasty_core::stmt::Type {
-        toasty_core::stmt::Type::String
-    }
-
-    fn load(value: toasty_core::stmt::Value) -> crate::Result<T> {
-        let json = match value {
-            toasty_core::stmt::Value::Null => std::borrow::Cow::Borrowed("null"),
-            v => std::borrow::Cow::Owned(<String as Load>::load(v)?),
-        };
-        serde_json::from_str(&json).map_err(|e| {
-            toasty_core::Error::from_args(format_args!("failed to deserialize JSON field: {e}"))
-        })
-    }
-}
