+++ crates/toasty/src/stmt.rs
@@ -30,6 +30,10 @@ pub use into_insert::IntoInsert;
 mod into_statement;
 pub use into_statement::IntoStatement;
 
+mod json;
+#[cfg(feature = "serde")]
+pub use json::Json;
+
 mod list;
 pub use list::List;
